refactor: movendo dtos de entrega non front
